more work and bug fixing towards cuda graphs

+++ common/buffers.py
@@ -389,7 +389,7 @@ class RolloutBuffer(BaseBuffer):
 
     def reset(self) -> None:
 
-        self.observations = np.zeros((self.max_steps, self.n_envs) + self.obs_shape, dtype=np.float16)
+        self.observations = np.zeros((self.max_steps, self.n_envs) + self.obs_shape, dtype=np.float32)
         self.actions = np.zeros((self.max_steps, self.n_envs, self.action_dim), dtype=np.float32)
         self.rewards = np.zeros((self.max_steps, self.n_envs), dtype=np.float32)
         self.returns = np.zeros((self.max_steps, self.n_envs), dtype=np.float32)
